requirements.txt
numpy
pandas
pyarrow
pyyaml
requests
lightgbm
scikit-learn
fastapi
uvicorn
pytest
